feature: #20 코드 리뷰 반영
