+++ src/main/kotlin/com/asap/asapbackend/batch/classroom/application/ClassroomScheduler.kt
@@ -1,5 +1,6 @@
 package com.asap.asapbackend.batch.classroom.application
 
+import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RestController
 
@@ -7,7 +8,7 @@ import org.springframework.web.bind.annotation.RestController
 class ClassroomScheduler(
         private val classroomService: ClassroomService
 ) {
-    //@Scheduled(cron = "0 0 4 1 3 ?") // 매년 3월 1일 04:00:00에 실행 //매년 학교 수가 달라져서 보류
+    @Scheduled(cron = "0 0 4 1 3 ?") // 매년 3월 1일 04:00:00에 실행
     @GetMapping("/addClassroom")
     fun addClassroom(){
         classroomService.addClassroom()
