+++ src/main/kotlin/com/asap/asapbackend/batch/classroom/application/ClassroomService.kt
@@ -1,12 +1,12 @@
 package com.asap.asapbackend.batch.classroom.application
 
-import com.asap.asapbackend.batch.classroom.service.ClassroomAppender
-import com.asap.asapbackend.batch.classroom.service.ClassroomReader
+import com.asap.asapbackend.domain.classroom.domain.service.ClassroomAppender
+import com.asap.asapbackend.batch.classroom.service.ClassroomOpenApiClient
 import org.springframework.stereotype.Service
 
 @Service
 class ClassroomService(
-        private val classroomReader: ClassroomReader,
+        private val classroomReader: ClassroomOpenApiClient,
         private val classroomAppender: ClassroomAppender
 ) {
     fun addClassroom() {
