+++ src/main/kotlin/com/asap/asapbackend/batch/school/application/ShoolScheduler.kt
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RestController
 class ShoolScheduler(
         private val schoolService: SchoolService
 ) {
-    //@Scheduled(cron = "0 0 4 1 3 ?") // 매년 3월 1일 04:00:00에 실행 //매년 학교 수가 달라져서 보류
+    @Scheduled(cron = "0 0 4 1 3 ?") // 매년 3월 1일 04:00:00에 실행
     @GetMapping("/addSchool")
     fun addSchool() {
         schoolService.addSchool()
