+++ src/main/kotlin/com/asap/asapbackend/domain/classroom/domain/repository/ClassroomRepository.kt
@@ -1,8 +1,10 @@
 package com.asap.asapbackend.domain.classroom.domain.repository
 
 import com.asap.asapbackend.domain.classroom.domain.model.Classroom
+import com.asap.asapbackend.domain.classroom.domain.model.Grade
+import com.asap.asapbackend.domain.school.domain.model.School
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface ClassroomRepository: JpaRepository<Classroom, Long> {
-    override fun deleteAll()
+    fun findBySchoolAndGradeAndClassNumber(school: School, grade: Grade, classNumber: String): Classroom?
 }
